services/agent/pyproject.toml
[project]
name = "serendip-agent"
version = "0.4.0"
description = "Serendip Bot — discovery agent and web crawler"
requires-python = ">=3.11"

dependencies = [
  # HTTP / crawling
  "httpx[http2]>=0.27.0",
  "beautifulsoup4>=4.12.3",
  "readability-lxml>=0.8.1",
  "robotexclusionrulesparser>=1.7.1",
  # AI providers
  "anthropic>=0.28.0",
  "google-genai>=1.0.0",
  "openai>=1.30.0",
  "ollama>=0.2.1",
  "azure-ai-inference>=1.0.0b5",
  "azure-core>=1.30.0",
  "azure-storage-blob>=12.14.0",
  # Queue / data
  "celery[redis]>=5.3.6",
  "gevent>=24.2.1",
  "flower>=2.0.1",
  "bullmq>=2.6.0",
  "redis[hiredis]>=5.0.6",
  "psycopg[binary]>=3.1.19",
  "pgvector>=0.3.2",
  # API shim
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.0",
  # ML / embedding
  "numpy>=1.26.4",
  # Config
  "pydantic>=2.7.1",
  "pydantic-settings>=2.3.0",
  "structlog>=24.2.0",
  "langdetect>=1.0.9",
  "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.4.6",
  "pytest>=8.2.2",
  "pytest-asyncio>=0.23.7",
  "pytest-httpx>=0.30.0",
  "mypy>=1.10.0",
]

image = [
  "transformers>=4.35.0",
  "torch>=2.0.0",
  "Pillow>=10.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
plugins = ["pydantic.mypy"]
disallow_untyped_decorators = false
warn_unused_ignores = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
